+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Build failing.
 - Glow does not register after reload.
 - Advice does not display.
+- Plugin offline.
 
 ## [7.2.0] - 2019-3-2
 ### Added